# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Vidya Socratic Tutor API
  description: |
    AI-powered Socratic tutoring API for STEM subjects and college essay coaching.
    
    Vidya uses the Socratic method to guide students toward understanding through
    carefully crafted questions — never giving direct answers or writing for the student.
    
    ## Authentication
    
    All API requests require a Bearer token (API key):
    ```
    Authorization: Bearer vk_live_xxxxxxxxxxxxxxxx
    ```
    
    API keys can be obtained by contacting the Vidya team.
    
    ## Rate Limits
    
    Rate limits are per-API-key and depend on your tier:
    - **Free**: 100 requests/minute — **hard limit** (429 when exceeded)
    - **Standard**: 500 requests/minute — **soft limit** (overage allowed and billed)
    - **Premium**: 1000 requests/minute — **soft limit** (overage allowed and billed)
    
    Standard and Premium tiers use a soft limit: requests above the plan
    limit are allowed but counted as overage and billed via Stripe metered
    billing. Free tier requests that exceed the limit receive a 429.
    
    Rate limit headers are included in every response:
    - `X-RateLimit-Limit`: Max requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Overage`: `true` when the request was allowed over the plan limit (Standard/Premium only)
    - `Retry-After`: Seconds to wait (only on 429 responses, Free tier only)
    
...
    
...
  version: 1.0.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/developer/status:
    get:
      summary: Get API key status
      description: |
        Returns the authenticated API key's configuration, status, usage stats,
        and available platform modules. Use this to check your key's tier,
        rate limit, session count, and what tutoring domains are available.
      tags:
        - Developer
      responses:
        '200':
          description: API key status and platform info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  key:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      prefix:
                        type: string
                      owner:
                        type: object
                        properties:
                          email:
                            type: string
                          name:
                            type: string
                      tier:
                        type: string
                        enum: [FREE, STANDARD, PREMIUM]
                      rateLimit:
                        type: integer
                      allowedOrigins:
                        type: array
                        items:
                          type: string
                      isActive:
                        type: boolean
                      lastUsedAt:
                        type: string
                        format: date-time
                        nullable: true
                      expiresAt:
                        type: string
                        format: date-time
                        nullable: true
                      totalRequests:
                        type: integer
                      totalSessions:
                        type: integer
                      billingEnabled:
                        type: boolean
                        description: Whether Stripe metered billing is linked to this key
                      createdAt:
                        type: string
                        format: date-time
                  platform:
                    type: object
                    properties:
                      availableModules:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            subjects:
                              type: array
                              items:
                                type: string
                      apiVersion:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/developer/usage:
    get:
      summary: Get usage analytics
      description: |
        Returns daily usage breakdown for the authenticated API key.
        Data is aggregated by date and endpoint.
      tags:
        - Developer
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
            maximum: 90
          description: Number of days to look back
        - name: endpoint
          in: query
          schema:
            type: string
          description: Filter by specific endpoint
      responses:
        '200':
          description: Daily usage data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  apiKeyId:
                    type: string
                  period:
                    type: object
                    properties:
                      days:
                        type: integer
                      from:
                        type: string
                  daily:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                        totalRequests:
                          type: integer
                        totalErrors:
                          type: integer
                        totalOverage:
                          type: integer
                          description: Requests that exceeded the plan rate limit (billed as overage)
                        endpoints:
                          type: object
                          additionalProperties:
                            type: integer
                  rawRecords:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/developer/usage/summary:
    get:
      summary: Get usage summary
      description: |
        Returns a quick summary of usage for the authenticated API key:
        requests today, this week, this month, top endpoints, and total sessions.
      tags:
        - Developer
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  apiKeyId:
                    type: string
                  summary:
                    type: object
                    properties:
                      today:
                        type: object
                        properties:
                          requests:
                            type: integer
                          errors:
                            type: integer
                          overage:
                            type: integer
                            description: Requests over plan limit (billed)
                      thisWeek:
                        type: object
                        properties:
                          requests:
                            type: integer
                          errors:
                            type: integer
                          overage:
                            type: integer
                      thisMonth:
                        type: object
                        properties:
                          requests:
                            type: integer
                          errors:
                            type: integer
                          overage:
                            type: integer
                      totalSessions:
                        type: integer
                      topEndpoints:
                        type: array
                        items:
                          type: object
                          properties:
                            endpoint:
                              type: string
                            requests:
                              type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the Vidya admin panel
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
...
  responses:
    Unauthorized:
      description: Authentication required or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error: Authentication required
            message: "Provide an API key via Authorization: Bearer <key>"
...